third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-019-expected.txt


FAIL CSS Test: @viewport constrained - used height can be zero assert_equals: expected 0 but got 1078
